sayma_rtm_drtio: replace sayma_rtm
